refactor: move address<=>line mappings into debugger
Will be needed by the breakpoint pane for file:line info.
