+++ lib/project/pep10.hpp
@@ -224,5 +224,4 @@ protected:
   QString _userAsmText = {}, _osAsmText = {};
   QString _userList = {}, _osList = {};
   QList<QPair<int, QString>> _errors = {}, _userListAnnotations = {}, _osListAnnotations = {};
-  helpers::AsmHelper::Lines2Addresses _userLines2Address = {}, _osLines2Address = {};
 };
